% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_modify_function.R
\name{energy_modify_function}
\alias{energy_modify_function}
\title{Energy Modify Function}
\usage{
energy_modify_function(energy_data)
}
\description{
This function allows you to preprocess easily.
abcabcabcabc
}
\examples{
energy_modify_function()
}
\keyword{data}
\keyword{energy}
\keyword{preprocessing}
